Simplifications and beautifications to fvwm and general GUI.
Added lines to .Xdefaults to handle grdc.

Took out a bunch of lines from .fvwm/config that weren't being used,
also added the pager to show desk and page setup (and added
keybindings similar to previous menu to bring it up and hide
it). Removed bubble popups from the pager too.

Added the MiniIcon for the terminal so it shows up everywhere. Added a
colorset for the pager, and changed the colors. Added automatic hot
keys for menus.

Fixed the annoying thing where my keyboard conf interferes with
pressing enter to select a menu item; made spacebar a menu exit so
that pulling up the window list with alt-tab can be exited more
easily. Made CS P iconify the window.

Slowed down the matrix screensaver in the matrix.c diff and in
mysuspend.

Forked grdc and added a Makefile to bin to make it.